Update README.md
Passionate Computer Science student at United International University in Dhaka, Bangladesh, with a fervent love for the field of machine learning. Eager to delve deeper into the realm of artificial intelligence and explore the limitless possibilities it offers. Driven by a vision to contribute to the advancements of this transformative technology, I am dedicated to continuous learning and growth. With a keen interest in exploring new horizons and pushing the boundaries of innovation, I aspire to become a Machine Learning Engineer. My journey is fueled by curiosity, determination, and a commitment to making a meaningful impact in the exciting era of AI.